등산하는 곳, 목적지 설정, 등산 진행 중 기록

+++ src/main/java/com/netj/deungchi/dto/mountain/MountainMapGetReqDto.java
@@ -15,7 +15,7 @@ public class MountainMapGetReqDto {
     private String content;
     private Boolean isShare;
     private String hikingDuration;
-    private Float hikingLength;
+    private Double hikingLength;
 
     public Record toRecordEntity(Member member, Mountain mountain, Course course) {
         return Record.builder()
